requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.2
Django==4.1.7
django-admin-interface==0.19.1
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-rest-knox==4.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pygraphviz==1.12
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.1.0
